+++ flumine/streams/basestream.py
@@ -3,6 +3,7 @@ import queue
 import logging
 import betfairlightweight
 from betfairlightweight import StreamListener
+from tenacity import wait_exponential
 
 logger = logging.getLogger(__name__)
 
@@ -11,6 +12,7 @@ class BaseStream(threading.Thread):
 
     LISTENER = StreamListener
     MAX_LATENCY = 0.5
+    RETRY_WAIT = wait_exponential(multiplier=1, min=2, max=20)
 
     def __init__(
         self,